requirements.txt
absl-py==2.1.0
APScheduler==3.11.0
astunparse==1.6.3
blinker==1.9.0
boto3==1.36.8
botocore==1.36.8
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.9
cssselect2==0.8.0
defusedxml==0.7.1
filelock==3.17.0
Flask==3.1.0
Flask-Mail==0.10.0
flatbuffers==25.1.24
fonttools==4.57.0
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
gTTS==2.5.4
h5py==3.12.1
html5lib==1.1
huggingface-hub==0.28.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
mariadb==1.1.12
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.1.0
protobuf==5.29.3
PyAudio==0.2.14
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
pyphen==0.17.2
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.3.0
requests==2.32.3
rich==13.9.4
s3transfer==0.11.2
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
SpeechRecognition==3.14.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==1.26.20
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==3.1.3
wkhtmltopdf==0.2
wrapt==1.17.2
zipp==3.21.0
zopfli==0.2.3.post1
